job_blueprint/requirements.txt
docling==2.36.0
python-dotenv==1.1.0
mistralai==1.8.1
langchain==0.3.25
langgraph==0.4.8
easyocr==1.7.2
openai==1.84.0
spacy-llm==0.7.3
fastapi==0.115.12
uvicorn==0.34.3
langchain-openai==0.3.21
elasticsearch==8.17.2
python-multipart